Reworked VEP behaviours - collected duplicate frequency stimulus code into a new abstract behaviour class shared for SSVEP and TVEP
