+++ Runtime/Scripts/Behaviors/SSVEPControllerBehavior.cs
@@ -1,11 +1,9 @@
-using System.Collections;
 using UnityEngine;
-using System;
 using BCIEssentials.Controllers;
 
 namespace BCIEssentials.ControllerBehaviors
 {
-    public class SSVEPControllerBehavior : WindowedControllerBehavior
+    public class SSVEPControllerBehavior : VEPControllerBehaviour
     {
         public override BCIBehaviorType BehaviorType => BCIBehaviorType.SSVEP;
         
@@ -17,31 +15,6 @@ namespace BCIEssentials.ControllerBehaviors
         [Tooltip("Calculated best-match achievable frequencies based on the application framerate [Hz]")]
         private float[] realFlashingFrequencies;
 
-        private int[] frames_on = new int[99];
-        private int[] frame_count = new int[99];
-        private float period;
-        private int[] frame_off_count = new int[99];
-        private int[] frame_on_count = new int[99];
-
-        public override void PopulateObjectList(SpoPopulationMethod populationMethod = SpoPopulationMethod.Tag)
-        {
-            base.PopulateObjectList(populationMethod);
-
-            realFlashingFrequencies = new float[_selectableSPOs.Count];
-
-            for (int i = 0; i < _selectableSPOs.Count; i++)
-            {
-                frames_on[i] = 0;
-                frame_count[i] = 0;
-                period = targetFrameRate / requestedFlashingFrequencies[i];
-                // could add duty cycle selection here, but for now we will just get a duty cycle as close to 0.5 as possible
-                frame_off_count[i] = (int)Math.Ceiling(period / 2);
-                frame_on_count[i] = (int)Math.Floor(period / 2);
-                realFlashingFrequencies[i] = (targetFrameRate / (float)(frame_off_count[i] + frame_on_count[i]));
-
-                Debug.Log($"frequency {i + 1} : {realFlashingFrequencies[i]}");
-            }
-        }
 
         protected override void SendWindowMarker(int trainingIndex = -1)
         => OutStream.PushSSVEPMarker(
@@ -49,51 +22,13 @@ namespace BCIEssentials.ControllerBehaviors
             realFlashingFrequencies, trainingIndex
         );
 
-        protected override IEnumerator OnStimulusRunBehavior()
-        {
-            // Add duty cycle
-            // Generate the flashing
-            for (int i = 0; i < _selectableSPOs.Count; i++)
-            {
-                frame_count[i]++;
-                if (frames_on[i] == 1)
-                {
-                    if (frame_count[i] >= frame_on_count[i])
-                    {
-                        // turn the cube off
-                        _selectableSPOs[i].StopStimulus();
-                        frames_on[i] = 0;
-                        frame_count[i] = 0;
-                    }
-                }
-                else
-                {
-                    if (frame_count[i] >= frame_off_count[i])
-                    {
-                        // turn the cube on
-                        _selectableSPOs[i].StartStimulus();
-                        frames_on[i] = 1;
-                        frame_count[i] = 0;
-                    }
-                }
-            }
 
-            yield return null;
+        protected override void InitializeFrequencies()
+        => realFlashingFrequencies = new float[_selectableSPOs.Count];
 
-            
-        }
-
-        protected override IEnumerator OnStimulusRunComplete()
-        {
-            foreach (var spo in _selectableSPOs)
-            {
-                if (spo != null)
-                {
-                    spo.StopStimulus();
-                }
-            }
-
-            yield return null;
-        }
+        protected override float GetRequestedFrequency(int index)
+        => requestedFlashingFrequencies[index];
+        protected override void SetRealFrequency(int index, float value)
+        => requestedFlashingFrequencies[index] = value;
     }
 }
