+++ Runtime/Scripts/Behaviors/TVEPControllerBehavior.cs
@@ -5,7 +5,7 @@ using BCIEssentials.Controllers;
 
 namespace BCIEssentials.ControllerBehaviors
 {
-    public class TVEPControllerBehavior : WindowedControllerBehavior
+    public class TVEPControllerBehavior : VEPControllerBehaviour
     {
         public override BCIBehaviorType BehaviorType => BCIBehaviorType.TVEP;
 
@@ -17,25 +17,6 @@ namespace BCIEssentials.ControllerBehaviors
         [Tooltip("Calculated best-match achievable frequency based on the application framerate [Hz]")]
         private float realFlashingFrequency;
 
-        private int[] frames_on = new int[99];
-        private int[] frame_count = new int[99];
-        private float period;
-        private int[] frame_off_count = new int[99];
-        private int[] frame_on_count = new int[99];
-
-        public override void PopulateObjectList(SpoPopulationMethod populationMethod = SpoPopulationMethod.Tag)
-        {
-            base.PopulateObjectList(populationMethod);
-            for (int i = 0; i < _selectableSPOs.Count; i++)
-            {
-                frames_on[i] = 0;
-                frame_count[i] = 0;
-                period = targetFrameRate / requestedFlashingFrequency;
-                frame_off_count[i] = (int)Math.Ceiling(period / 2);
-                frame_on_count[i] = (int)Math.Floor(period / 2);
-                realFlashingFrequency = (targetFrameRate / (float)(frame_off_count[i] + frame_on_count[i]));
-            }
-        }
 
         protected override void SendWindowMarker(int trainingIndex = -1)
         => OutStream.PushTVEPMarker(
@@ -43,48 +24,10 @@ namespace BCIEssentials.ControllerBehaviors
             new[] {realFlashingFrequency}, trainingIndex
         );
 
-        protected override IEnumerator OnStimulusRunBehavior()
-        {
-            // Add duty cycle
-            // Generate the flashing
-            for (int i = 0; i < _selectableSPOs.Count; i++)
-            {
-                frame_count[i]++;
-                if (frames_on[i] == 1)
-                {
-                    if (frame_count[i] >= frame_on_count[i])
-                    {
-                        // turn the stimulus off
-                        _selectableSPOs[i].StopStimulus();
-                        frames_on[i] = 0;
-                        frame_count[i] = 0;
-                    }
-                }
-                else
-                {
-                    if (frame_count[i] >= frame_off_count[i])
-                    {
-                        // turn the stimulus on
-                        _selectableSPOs[i].StartStimulus();
-                        frames_on[i] = 1;
-                        frame_count[i] = 0;
-                    }
-                }
-            }
-            yield return null;
-        }
 
-        protected override IEnumerator OnStimulusRunComplete()
-        {
-            foreach (var spo in _selectableSPOs)
-            {
-                if (spo != null)
-                {
-                    spo.StopStimulus();
-                }
-            }
-
-            yield return null;
-        }
+        protected override float GetRequestedFrequency(int index)
+        => requestedFlashingFrequency;
+        protected override void SetRealFrequency(int index, float value)
+        => realFlashingFrequency = value;
     }
 }
